Roll back implicit cast
What: Roll back the implicit cast, isolating this PR to just adding unit tests
Why: Change was not covered by unit tests, and didn't fit the rest of the PR scope

+++ src/main/java/org/apache/commons/imaging/palette/ColorGroup.java
@@ -118,10 +118,10 @@ final class ColorGroup {
 
         for (final ColorCount color : colorCounts) {
             countTotal += color.count;
-            alphaTotal += (long) color.count * color.alpha;
-            redTotal += (long) color.count * color.red;
-            greenTotal += (long) color.count * color.green;
-            blueTotal += (long) color.count * color.blue;
+            alphaTotal += color.count * color.alpha;
+            redTotal += color.count * color.red;
+            greenTotal += color.count * color.green;
+            blueTotal += color.count * color.blue;
         }
 
         final int alpha = ignoreAlpha ? 0xff : (int) Math.round((double) alphaTotal / countTotal);
